Tweak: Conversion Center - refactor base class
